refactor(model): move data classes to model dir
